support addonDeploymentConfig.ProxyConfig for managed cluster components

+++ operators/multiclusterobservability/controllers/multiclusterobservability/multiclusterobservability_controller_test.go
@@ -59,8 +59,10 @@ func setupTest(t *testing.T) func() {
 	os.Setenv("TEMPLATES_PATH", testManifestsPath)
 	templates.ResetTemplates()
 	//clean up the manifest path if left over from previous test
-	if err = os.Remove(testManifestsPath); err != nil {
-		t.Log(fmt.Sprintf("Failed to delete symbollink(%s) for the test manifests: (%v)", testManifestsPath, err))
+	if fi, err := os.Lstat(testManifestsPath); err == nil && fi.Mode()&os.ModeSymlink != 0 {
+		if err = os.Remove(testManifestsPath); err != nil {
+			t.Logf(fmt.Sprintf("Failed to delete symlink(%s) for the test manifests: (%v)", testManifestsPath, err))
+		}
 	}
 	err = os.Symlink(manifestsPath, testManifestsPath)
 	if err != nil {
@@ -71,7 +73,7 @@ func setupTest(t *testing.T) func() {
 	return func() {
 		t.Log("begin teardownTest")
 		if err = os.Remove(testManifestsPath); err != nil {
-			t.Log(fmt.Sprintf("Failed to delete symbollink(%s) for the test manifests: (%v)", testManifestsPath, err))
+			t.Logf(fmt.Sprintf("Failed to delete symbollink(%s) for the test manifests: (%v)", testManifestsPath, err))
 		}
 		os.Remove(path.Join(wd, "../../tests"))
 		os.Unsetenv("TEMPLATES_PATH")
